feat: refactor app module to chef entity in gateway
