+++ gateway/src/entity/chef.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
+import { AppController } from './chef.controller';
+import { AppService } from './chef.service';
 
 describe('AppController', () => {
   let appController: AppController;
