+++ gateway/src/entity/chef.service.ts
@@ -1,11 +1,11 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { ClientProxy } from '@nestjs/microservices';
 import { lastValueFrom, Observable } from 'rxjs';
-import { ChefModel } from './model/chef.model';
-import { ChefLoginModel } from '../../chef-service/src/model/chef.model.dto';
+import { ChefModel } from '../model/chef.model';
+import { ChefLoginModel } from '../../../chef-service/src/model/chef.model.dto';
 
 @Injectable()
-export class AppService {
+export class ChefService {
   constructor(
     @Inject('CHEF_SERVICE') private readonly chefClient: ClientProxy,
   ) {}
